requirements.txt
requests=2.31.0
torch=2.3.0
transformers=4.40.2
pillow=10.3.0
streamlit=1.34.0
python-box=7.1.1